requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.4.0
pylint==2.2.2
PyYAML==3.13
requests==2.21.0
six==1.12.0
transmissionrpc==0.11
urllib3==1.24.1
wrapt==1.10.11
